fix import and declare order
