+++ python/sglang/srt/layers/quantization/kv_cache.py
@@ -8,7 +8,7 @@ from sglang.srt.layers.quantization.base_config import (
     QuantizationConfig,
     QuantizeMethodBase,
 )
-from sglang.srt.layers.quantization.utils import is_fp8_fnuz
+from sglang.srt.layers.quantization.fp8_kernel import is_fp8_fnuz
 from sglang.srt.layers.radix_attention import RadixAttention
 
 logger = logging.getLogger(__name__)
